add guard to correct routes
