+++ src/applications/mhv/medications/routes.jsx
@@ -2,7 +2,6 @@ import React from 'react';
 import { Switch, Route } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import PageNotFound from '@department-of-veterans-affairs/platform-site-wide/PageNotFound';
-import { useMyHealthAccessGuard } from '~/platform/mhv/hooks/useMyHealthAccessGuard';
 import App from './containers/App';
 import PrescriptionDetails from './containers/PrescriptionDetails';
 import RxBreadcrumbs from './containers/RxBreadcrumbs';
@@ -10,6 +9,9 @@ import Prescriptions from './containers/Prescriptions';
 import LandingPage from './containers/LandingPage';
 import RefillPrescriptions from './containers/RefillPrescriptions';
 
+/**
+ * Route that wraps its children within the application component.
+ */
 const AppRoute = ({ children, ...rest }) => {
   return (
     <Route {...rest}>
@@ -25,50 +27,40 @@ AppRoute.propTypes = {
   children: PropTypes.object,
 };
 
-const AccessGuardWrapper = ({ children }) => {
-  const redirectToMyHealth = useMyHealthAccessGuard();
-  if (redirectToMyHealth) {
-    return redirectToMyHealth;
-  }
-  return children;
-};
-
 const routes = (
   <div className="vads-l-grid-container">
     <div className="main-content vads-l-col--12 medium-screen:vads-l-col--8 medium-screen:vads-u-margin-left--neg2 vads-u-max-width--100">
-      <AccessGuardWrapper>
-        <Switch>
-          <AppRoute exact path={['/about', '/about/*']} key="LandingPage">
-            <LandingPage />
-          </AppRoute>
-          <AppRoute exact path={['/refill']} key="RefillPage">
-            <div>
-              <RefillPrescriptions />
-              <div className="no-print">
-                <va-back-to-top />
-              </div>
+      <Switch>
+        <AppRoute exact path={['/about', '/about/*']} key="LandingPage">
+          <LandingPage />
+        </AppRoute>
+        <AppRoute exact path={['/refill']} key="RefillPage">
+          <div>
+            <RefillPrescriptions />
+            <div className="no-print">
+              <va-back-to-top />
             </div>
-          </AppRoute>
-          <AppRoute exact path={['/', '/:page']} key="App">
-            <div>
-              <Prescriptions />
-              <div className="no-print">
-                <va-back-to-top />
-              </div>
+          </div>
+        </AppRoute>
+        <AppRoute exact path={['/', '/:page']} key="App">
+          <div>
+            <Prescriptions />
+            <div className="no-print">
+              <va-back-to-top />
             </div>
-          </AppRoute>
-          <AppRoute
-            exact
-            path="/prescription/:prescriptionId"
-            key="prescriptionDetails"
-          >
-            <PrescriptionDetails />
-          </AppRoute>
-          <Route>
-            <PageNotFound />
-          </Route>
-        </Switch>
-      </AccessGuardWrapper>
+          </div>
+        </AppRoute>
+        <AppRoute
+          exact
+          path="/prescription/:prescriptionId"
+          key="prescriptionDetails"
+        >
+          <PrescriptionDetails />
+        </AppRoute>
+        <Route>
+          <PageNotFound />
+        </Route>
+      </Switch>
     </div>
   </div>
 );
